attempting move of object lookup to UTILS.render
